Extract exchange rate display to component
